Add service functions to handle data manipulation

+++ src/services/trend.service.ts
@@ -1,12 +1,19 @@
-import { INITIAL_NEWS } from "../config/db.config";
+import Feed from "../models/feed.model";
 import Trend, { ITrend } from "../models/trends.model";
 export class TrendService {
   public async getTrends() {
-    return Trend.find().limit(INITIAL_NEWS);
+    return Trend.find();
   }
 
-  public async createTrend(trend: ITrend) {
+  public async addTrend(trend: ITrend) {
     const newTrend = new Trend(trend);
+    const feed = await this.getFeed();
+    if (!feed) {
+      return;
+    }
+
+    feed.trends.push(newTrend.id);
+    await feed.save();
     return newTrend.save();
   }
 
@@ -17,6 +24,27 @@ export class TrendService {
   }
 
   public async deleteTrend(id: string) {
-    return Trend.findByIdAndDelete(id);
+    const feed = await this.getFeed();
+    if (!feed) {
+      return;
+    }
+    await Trend.findByIdAndDelete(id);
+    feed.trends = feed.trends.filter((t) => t !== id);
+    return feed?.save();
+  }
+
+  private async getFeed() {
+    const startOfToday = new Date();
+    startOfToday.setHours(0, 0, 0, 0);
+    const endOfToday = new Date();
+    endOfToday.setHours(23, 59, 59, 999);
+    const feed = Feed.findOne({
+      date: {
+        $gt: startOfToday,
+        $lt: endOfToday,
+      },
+    }).populate("trends");
+
+    return feed;
   }
 }
